# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ImageAnalysisInput {
  /** Base64 encoded image data (with or without data URI prefix) */
  imageBase64: string;
}

/**
 * @nullable
 */
export type SpiderAnalysisResultDangerLevel = typeof SpiderAnalysisResultDangerLevel[keyof typeof SpiderAnalysisResultDangerLevel] | null;


export const SpiderAnalysisResultDangerLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface SpiderAnalysisResult {
  id: number;
  spiderName: string;
  scientificName: string;
  /** Confidence percentage 0-100 */
  confidence: number;
  isDangerous: boolean;
  /** @nullable */
  dangerLevel?: SpiderAnalysisResultDangerLevel;
  habitat: string;
  diet: string;
  lifespan: string;
  facts: string[];
  safetyInfo: string;
  relatedSpecies: string[];
  /** @nullable */
  imageBase64?: string | null;
  analyzedAt: string;
}

export type BiteAnalysisResultDangerLevel = typeof BiteAnalysisResultDangerLevel[keyof typeof BiteAnalysisResultDangerLevel];


export const BiteAnalysisResultDangerLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface BiteAnalysisResult {
  id: number;
  possibleSpider: string;
  dangerLevel: BiteAnalysisResultDangerLevel;
  symptoms: string[];
  safetyTips: string[];
  recommendation: string;
  /** @nullable */
  imageBase64?: string | null;
  analyzedAt: string;
}

export type ScanHistoryType = typeof ScanHistoryType[keyof typeof ScanHistoryType];


export const ScanHistoryType = {
  spider: 'spider',
  bite: 'bite',
} as const;

export interface ScanHistory {
  id: number;
  type: ScanHistoryType;
  name: string;
  isDangerous: boolean;
  /** @nullable */
  confidence: number | null;
  /** @nullable */
  imageBase64?: string | null;
  createdAt: string;
}

export interface ScanStats {
  totalScans: number;
  dangerousFound: number;
  safeFound: number;
  spiderScans: number;
  biteScans: number;
}

export type GetHistoryParams = {
search?: string;
filter?: GetHistoryFilter;
};

export type GetHistoryFilter = typeof GetHistoryFilter[keyof typeof GetHistoryFilter];


export const GetHistoryFilter = {
  all: 'all',
  dangerous: 'dangerous',
  safe: 'safe',
  today: 'today',
  this_week: 'this_week',
} as const;