# lib/api-zod/src/generated/types/biteAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BiteAnalysisResultDangerLevel } from './biteAnalysisResultDangerLevel';

export interface BiteAnalysisResult {
  id: number;
  possibleSpider: string;
  dangerLevel: BiteAnalysisResultDangerLevel;
  symptoms: string[];
  safetyTips: string[];
  recommendation: string;
  /** @nullable */
  imageBase64?: string | null;
  analyzedAt: Date;
}