# lib/api-zod/src/generated/types/imageAnalysisInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageAnalysisInput {
  /** Base64 encoded image data (with or without data URI prefix) */
  imageBase64: string;
}