# lib/api-zod/src/generated/types/scanHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanHistoryType } from './scanHistoryType';

export interface ScanHistory {
  id: number;
  type: ScanHistoryType;
  name: string;
  isDangerous: boolean;
  /** @nullable */
  confidence: number | null;
  /** @nullable */
  imageBase64?: string | null;
  createdAt: Date;
}