# lib/api-zod/src/generated/types/scanStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanStats {
  totalScans: number;
  dangerousFound: number;
  safeFound: number;
  spiderScans: number;
  biteScans: number;
}