# lib/api-zod/src/generated/types/spiderAnalysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SpiderVision AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SpiderAnalysisResultDangerLevel } from './spiderAnalysisResultDangerLevel';

export interface SpiderAnalysisResult {
  id: number;
  spiderName: string;
  scientificName: string;
  /** Confidence percentage 0-100 */
  confidence: number;
  isDangerous: boolean;
  /** @nullable */
  dangerLevel?: SpiderAnalysisResultDangerLevel;
  habitat: string;
  diet: string;
  lifespan: string;
  facts: string[];
  safetyInfo: string;
  relatedSpecies: string[];
  /** @nullable */
  imageBase64?: string | null;
  analyzedAt: Date;
}